requirements.txt
flask>=3.1.0
requests>=2.32.3
dnspython>=2.7.0
python-whois>=0.9.5
beautifulsoup4>=4.12.3
python-Wappalyzer>=0.3.1
tldextract>=5.1.3
lxml>=5.3.0
